# 原型/lumina-api-openapi.yaml
openapi: 3.1.0
info:
  title: Lumina 墨光 · API
  version: 1.5.0
  description: |
    Lumina（墨光）教育应用 RESTful API，覆盖用户、课程、作业、成绩、直播、协作、AI 等模块。

    ## 特性
    - JWT 双令牌认证（Access Token 15min + Refresh Token 7d）
    - Cursor-based 分页
    - RFC 7807 错误格式
    - SSE 流式响应（AI 接口）
    - 直播课堂：REST + after_id 轮询（聊天/举手/答题/点名）· HLS 流媒体（LIVE_STREAM_* 可插拔，未接入返回 mock://）
    - 小组协作：课程小组 / 成员 / 项目 / 看板拖拽 / 共享文件 / 组内讨论（D-02）
    - 账户补全：自助注册（POST /auth/register）+ 消息通知中心（/notifications/my · D-03）

  contact:
    name: Lumina Team
    url: https://github.com/yuan50697105/lumina-edu
  license:
    name: MIT
servers:
  - url: https://api.lumina.edu/api/v1
    description: 生产环境
  - url: https://api-staging.lumina.edu/api/v1
    description: 预发布环境
  - url: http://localhost:8080/api/v1
    description: 本地开发
...
security:
  - BearerAuth: []

# ─── 路径 ───
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: 登录获取令牌
      description: 使用账号密码登录，返回 Access Token 和 Refresh Token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  data:
                    $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [auth]
      summary: 刷新令牌
      description: 使用 Refresh Token 获取新的 Access Token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIs..."
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  data:
                    $ref: '#/components/schemas/TokenResponse'
...
  /auth/password:
    post:
      tags: [auth]
      summary: 修改密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [old_password, new_password]
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/register:
    post:
      tags: [auth]
      summary: 自助注册（D-03）
      description: 注册学生/教师账号，成功即自动登录返回令牌；邮箱/学号冲突返回 409
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功，返回令牌
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  data:
                    $ref: '#/components/schemas/TokenResponse'
        '409':
          description: 邮箱或学号已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 参数校验失败（角色非法 / 密码过短）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ═══════ USERS ═══════
...
  /notifications/my:
    get:
      tags: [notif]
      summary: 我的通知列表
      description: 仅本人可见；支持 limit / unread_only 过滤
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: unread_only
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: 通知列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationOut'
...
  /notifications/my/{notification_id}/read:
    post:
      tags: [notif]
      summary: 单条已读
      description: 非本人通知返回 404
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 已标记已读
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Access Token（15 分钟有效）"

  # ─── 通用参数 ───
...
  responses:
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: "success"
...
    Error:
      type: object
      description: "RFC 7807 Problem Details"
      properties:
        type:
          type: string
          format: uri
          example: "https://api.lumina.edu/errors/unauthorized"
        status:
          type: integer
          example: 401
        title:
          type: string
          example: "Unauthorized"
        detail:
          type: string
          example: "Access token expired"
        code:
          type: string
          example: "AUTH_TOKEN_EXPIRED"
        request_id:
          type: string

    # --- Auth ---
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "linqing"
        password:
          type: string
          format: password
        device:
          type: string
          enum: [web, mobile, desktop]
          default: web
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: "秒"
          example: 900
        token_type:
          type: string
          example: "Bearer"
        user:
          $ref: '#/components/schemas/User'

    # --- User ---
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        student_id:
          type: string
          description: "学号"
          example: "2023010001"
        name:
          type: string
          example: "林清"
        email:
          type: string
          format: email
        role:
          type: string
          enum: [student, teacher, admin]
        department:
          type: string
        grade:
          type: string
          description: "入学年级"
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        created_at:
          type: string
          format: date-time

    # --- Course ---
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: 至少 8 位
        student_id:
          type: string
          description: 学号/工号（可选）
        role:
          type: string
          enum: [student, teacher]
          default: student
        device:
          type: string
          enum: [web, mobile, desktop]
          default: web
    NotificationOut:
      type: object
      required: [id, user_id, type, title, is_read, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        type:
          type: string
          description: 通知类型（welcome / assignment_graded / live_call）
        title:
          type: string
        content:
          type: string
          nullable: true
        ref_type:
          type: string
          nullable: true
          description: 关联对象类型（course / live_room / assignment）
        ref_id:
          type: string
          nullable: true
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
...